refactor payments

+++ internal/usecase/payment/delete_payment.go
@@ -3,10 +3,13 @@ package payment
 import (
 	"context"
 	"errors"
+	"fmt"
 
+	"github.com/rober0xf/notifier/internal/domain/entity"
 	domainErr "github.com/rober0xf/notifier/internal/domain/errors"
 	"github.com/rober0xf/notifier/internal/domain/repository"
 	repoErr "github.com/rober0xf/notifier/internal/infraestructure/errors"
+	authErr "github.com/rober0xf/notifier/pkg/auth"
 )
 
 type DeletePaymentUseCase struct {
@@ -19,16 +22,30 @@ func NewDeletePaymentUseCase(paymentRepo repository.PaymentRepository) *DeletePa
 	}
 }
 
-func (uc *DeletePaymentUseCase) Execute(ctx context.Context, id int) error {
+func (uc *DeletePaymentUseCase) Execute(ctx context.Context, id, userID int, userRole entity.Role) error {
 	if id <= 0 {
 		return domainErr.ErrInvalidPaymentData
+	}
+
+	payment, err := uc.paymentRepo.GetPaymentByID(ctx, id)
+	if err != nil {
+		if errors.Is(err, repoErr.ErrNotFound) {
+			return domainErr.ErrPaymentNotFound
+		}
+
+		return fmt.Errorf("failed getting payment by id: %w", err)
+	}
+
+	if payment.UserID != userID && userRole != entity.RoleAdmin {
+		return authErr.ErrForbidden
 	}
 
 	if err := uc.paymentRepo.DeletePayment(ctx, id); err != nil {
 		if errors.Is(err, repoErr.ErrNotFound) {
 			return domainErr.ErrPaymentNotFound
 		}
-		return err
+
+		return fmt.Errorf("DeletePaymentUC.Execute failed to delete payment: %w", err)
 	}
 
 	return nil
